refactor style sheets and remove logo.svg
