+++ client/src/App.js
@@ -7,7 +7,7 @@ import { Register } from "./pages/Register";
 import { Home } from "./pages/Home";
 import { Survey } from "./pages/Survey";
 import { Search } from "./pages/Search";
-import "./App.css";
+import "./styles/App.css";
 import { Navbar } from "./components/Navbar";
 import { Footer } from "./components/Footer";
