+++ client/src/pages/Login.jsx
@@ -3,6 +3,8 @@ import { useHistory } from "react-router-dom";
 import { loginUser } from "../utils/auth";
 import { Container, Form } from "react-bootstrap";
 import { Button } from "../components/Button";
+import "../styles/Form.css";
+
 
 export const Login = () => {
   const history = useHistory();
