+++ client/src/styles/Home.css
@@ -1,0 +1,16 @@
+.description {
+    left: 50%;
+    position: absolute;
+    top: 45%;
+    transform: translate(-50%, -50%);
+    text-align: center;
+}
+
+.description h1 {
+    color: #6ab446;
+}
+
+.description p {
+    font-size: 1.3rem;
+    line-height: 1.5;
+}
